dot2dot/requirements.txt
matplotlib==3.9.4
numba==0.60.0
numpy==2.0.0
opencv_python==4.10.0.84
Pillow==11.0.0
scikit-image
